+++ internal/handler/envoyextauth/grpcv3/middleware/errorhandler/interceptor_test.go
@@ -32,7 +32,7 @@ import (
 	"google.golang.org/grpc/credentials/insecure"
 	"google.golang.org/grpc/test/bufconn"
 
-	"github.com/dadrus/heimdall/internal/handler/envoyextauth/grpcv3/mocks"
+	"github.com/dadrus/heimdall/internal/handler/envoyextauth/grpcv3/middleware/mocks"
 	"github.com/dadrus/heimdall/internal/heimdall"
 )
